Add files via upload
Add this text to your .bashrc, save it, restart the terminal, run "mutespotifyads" in terminal to see if it works with the desktop Spotify application. Finally, if it works, you can try "adfree" in the terminal to run this command in a detached tmux session (so you don't need to have a terminal running in the foreground the whole time).